opensource/sdk-python/pyproject.toml
# pyproject.toml

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cord-sdk_server"
version = "0.0.1"
description = "The Cord Server SDK for Python"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
]
keywords = ["cord", "sdk", "server", "cord-sdk"]
dependencies = [ "PyJWT" ]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/getcord/sdk-python"
